Move Anomaly Scanner DoAfter code to shared
