+++ Content.Shared/Anomaly/SharedAnomalyScannerSystem.cs
@@ -1,4 +1,7 @@
-﻿using Content.Shared.Popups;
+﻿using Content.Shared.Anomaly.Components;
+using Content.Shared.DoAfter;
+using Content.Shared.Interaction;
+using Content.Shared.Popups;
 using Robust.Shared.Audio.Systems;
 using Robust.Shared.Timing;
 
@@ -10,4 +13,49 @@ public abstract class SharedAnomalyScannerSystem : EntitySystem
     [Dependency] protected readonly SharedAudioSystem Audio = default!;
     [Dependency] protected readonly IGameTiming Timing = default!;
     [Dependency] protected readonly SharedAppearanceSystem Appearance = default!;
+    [Dependency] private readonly SharedDoAfterSystem _doAfter = default!;
+    [Dependency] protected readonly SharedUserInterfaceSystem UI = default!;
+
+
+    public override void Initialize()
+    {
+        base.Initialize();
+
+        SubscribeLocalEvent<AnomalyScannerComponent, AfterInteractEvent>(OnScannerAfterInteract);
+    }
+
+    private void OnScannerAfterInteract(EntityUid uid, AnomalyScannerComponent component, AfterInteractEvent args)
+    {
+        if (args.Target is not { } target)
+            return;
+        if (!HasComp<AnomalyComponent>(target))
+            return;
+        if (!args.CanReach)
+            return;
+
+        _doAfter.TryStartDoAfter(new DoAfterArgs(EntityManager,
+            args.User,
+            component.ScanDoAfterDuration,
+            new ScannerDoAfterEvent(),
+            uid,
+            target: target,
+            used: uid)
+        {
+            DistanceThreshold = 2f
+        });
+    }
+
+    protected virtual void OnDoAfter(EntityUid uid, AnomalyScannerComponent component, DoAfterEvent args)
+    {
+        if (args.Cancelled || args.Handled || args.Args.Target == null)
+            return;
+
+        Audio.PlayPvs(component.CompleteSound, uid);
+        Popup.PopupEntity(Loc.GetString("anomaly-scanner-component-scan-complete"), uid);
+
+        UI.OpenUi(uid, AnomalyScannerUiKey.Key, args.User);
+
+        args.Handled = true;
+    }
+
 }
